+++ projects/angular-esri-components/src/lib/core/esri-module-provider.ts
@@ -1,17 +1,17 @@
 import { Injectable, ElementRef, EventEmitter } from '@angular/core';
-import { EsriLoaderService } from 'angular-esri-loader';
+import { loadModules, loadScript } from 'esri-loader';
 import { environment } from '../../environments/environment';
 
 @Injectable()
 export class EsriModuleProvider {
-  constructor(private esriLoader: EsriLoaderService) { }
+  constructor() {}
 
   require(moduleNames: string[]): Promise<any[]> {
-    return this.esriLoader.load({
-        // the specific version of the API that is to be used
-        url: environment.arcGisScriptUrl
-      }).then( () => {
-              return this.esriLoader.loadModules(moduleNames);
-      });
+    return loadScript({
+      // the specific version of the API that is to be used
+      url: environment.arcGisScriptUrl
+    }).then(() => {
+      return loadModules(moduleNames);
+    });
   }
 }